fix deploy script
Now it doesn't get confused by `master` not being the default branch.
Also, it uses the ssh github url. This means that you need to have ssh
access to the repo and that you won't be asked for your username and
password on every deploy. You have to delete ~/.servant-www once for
that to work.